+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## 1.0.0
-- Initial release
+- Initial release.
+
+## 1.0.1
+- Added `trailing-whitespace-fixer.allowWhitespaceOnlyLines` setting to allow lines that are only whitespace.
+- Fix highlighting suppression on last line of file.